refactor(rust): move `list::verify_pids` to `util`
