+++ implementations/rust/ockam/ockam_command/src/node/list.rs
@@ -1,11 +1,6 @@
-use std::time::Duration;
-
-use crate::util::{self, api, connect_to, exitcode};
-use crate::{node::show::query_status, CommandGlobalOpts, OckamConfig};
+use crate::util::{connect_to, exitcode, verify_pids};
+use crate::{node::show::query_status, CommandGlobalOpts};
 use clap::Args;
-use crossbeam_channel::{bounded, Sender};
-use ockam::{Context, Route};
-use ockam_api::nodes::NODEMANAGER_ADDR;
 
 #[derive(Clone, Debug, Args)]
 pub struct ListCommand {}
@@ -43,55 +38,4 @@ impl ListCommand {
                 )
             });
     }
-}
-
-// TODO: move to utils directory
-fn verify_pids(cfg: &OckamConfig, nodes: Vec<String>) {
-    for node_name in nodes {
-        let node_cfg = cfg.get_node(&node_name).unwrap();
-
-        let (tx, rx) = bounded(1);
-
-        connect_to(node_cfg.port, tx, query_pid);
-        let verified_pid = rx.recv().unwrap();
-
-        if node_cfg.pid != verified_pid {
-            if let Err(e) = cfg.update_pid(&node_name, verified_pid) {
-                eprintln!("failed to update pid for node {}: {}", node_name, e);
-                std::process::exit(exitcode::IOERR);
-            }
-        }
-    }
-
-    if cfg.atomic_update().run().is_err() {
-        eprintln!("failed to update PID information in config!");
-        std::process::exit(exitcode::IOERR);
-    }
-}
-
-pub async fn query_pid(
-    mut ctx: Context,
-    tx: Sender<Option<i32>>,
-    mut base_route: Route,
-) -> anyhow::Result<()> {
-    ctx.send(
-        base_route.modify().append(NODEMANAGER_ADDR),
-        api::query_status()?,
-    )
-    .await?;
-
-    let resp = match ctx
-        .receive_duration_timeout::<Vec<u8>>(Duration::from_millis(200))
-        .await
-    {
-        Ok(r) => r.take().body(),
-        Err(_) => {
-            tx.send(None).unwrap();
-            return util::stop_node(ctx).await;
-        }
-    };
-
-    let status = api::parse_status(&resp)?;
-    tx.send(Some(status.pid)).unwrap();
-    util::stop_node(ctx).await
 }
