+++ implementations/rust/ockam/ockam_command/src/util/mod.rs
@@ -6,6 +6,7 @@ use std::{
 };
 
 use anyhow::{anyhow, Context as _, Result};
+use crossbeam_channel::{bounded, Sender};
 use minicbor::{Decode, Decoder, Encode};
 use tracing::{debug, error, trace};
 use tracing_subscriber::prelude::*;
@@ -417,4 +418,54 @@ pub fn bind_to_port_check(address: &SocketAddr) -> bool {
     let port = address.port();
     let ip = address.ip();
     std::net::TcpListener::bind((ip, port)).is_ok()
+}
+
+pub fn verify_pids(cfg: &OckamConfig, nodes: Vec<String>) {
+    for node_name in nodes {
+        let node_cfg = cfg.get_node(&node_name).unwrap();
+
+        let (tx, rx) = bounded(1);
+
+        connect_to(node_cfg.port, tx, query_pid);
+        let verified_pid = rx.recv().unwrap();
+
+        if node_cfg.pid != verified_pid {
+            if let Err(e) = cfg.update_pid(&node_name, verified_pid) {
+                eprintln!("failed to update pid for node {}: {}", node_name, e);
+                std::process::exit(exitcode::IOERR);
+            }
+        }
+    }
+
+    if cfg.atomic_update().run().is_err() {
+        eprintln!("failed to update PID information in config!");
+        std::process::exit(exitcode::IOERR);
+    }
+}
+
+pub async fn query_pid(
+    mut ctx: Context,
+    tx: Sender<Option<i32>>,
+    mut base_route: Route,
+) -> anyhow::Result<()> {
+    ctx.send(
+        base_route.modify().append(NODEMANAGER_ADDR),
+        api::query_status()?,
+    )
+    .await?;
+
+    let resp = match ctx
+        .receive_duration_timeout::<Vec<u8>>(Duration::from_millis(200))
+        .await
+    {
+        Ok(r) => r.take().body(),
+        Err(_) => {
+            tx.send(None).unwrap();
+            return stop_node(ctx).await;
+        }
+    };
+
+    let status = api::parse_status(&resp)?;
+    tx.send(Some(status.pid)).unwrap();
+    stop_node(ctx).await
 }
